ai-moderator/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "benchradar-ai-moderator"
version = "0.0.1"
requires-python = ">=3.10"
dependencies = [
  "transformers>=4.40",
  "torch",
  "Pillow>=10",
]

[project.scripts]
bench-prob = "ai_moderator.cli:main"

[tool.setuptools]
packages = ["ai_moderator"]
